Add more text output to service test pause_on_start_then_step_test
This test has timed out on the bots, but it isn't clear when exactly
it 'stops responding'.

This CL adds a few more lines of output in an attempt to make it
more clear (on any new timeout) where it stops responding.

See https://github.com/dart-lang/sdk/issues/29568 for example timeout.

R=kmillikin@google.com

Review-Url: https://codereview.chromium.org/2861673008 .