Update expectations for instance creations to match front_end.
Front_end wants to annotate instance creations like this:

    new /*@typeArgs=Foo*/ bar();

Whereas the existing tests use this style:

    /*@typeArgs=Foo*/ new bar();

It's easier to modify the analyzer code to use the front_end style
than to update the front_end to use the analyzer style, and besides, I
think the front_end style makes more sense.

I've also added a boolean to front_end_inference_test.dart to allow
expectations to be updated more easily in the future.

R=scheglov@google.com

Review-Url: https://codereview.chromium.org/2867113002 .