Remove generic parameters from type inference classes.
Previously, these classes were generic because we anticapted re-using
them with analyzer AST objects in order to do analyzer type inference.
But we've decided on a new technique for doing analyzer type
inference: infer at the kernel level and then transfer the inference
data to analyzer using the TypeInferenceListener.

So the generic parameters are no longer needed.

R=scheglov@google.com

Review-Url: https://codereview.chromium.org/2885873002 .