Hit breakpoints in single line closures.
Before this change, breakpoints on lines like the following were not hit:

var singleLineClosure = (int a, int b) => return a + b;

The reason was that breakpoints were only resolved if their start token
was contained within a function. Since that is not true for closure vars
like in the above example, the breakpoints were never getting resolved.
This change avoids such a problem by resolving breakpoints if the token
range of the breakpoint overlaps the token range of a function.

Fixes #29736

R=asiva@google.com

Review-Url: https://codereview.chromium.org/2912763005 .