Fix ParameterElementImpl.hasImplicitType for function-typed parameters.
UnlinkedParam.type serves a dual role: for ordinary parameters it is
the declared type; for function-typed parameters it is the declared
return type.  So to determine whether a parameter has an implicit type
it is not sufficient to check whether UnlinkedParam.type is `null`.
We also need to check whether the parameter is function-typed.

R=brianwilkerson@google.com

Review-Url: https://codereview.chromium.org/2947703002 .