VM-codegen: For polymorphic methods don't expand Cids at call sites
When dispatching polymorphic calls we try to recognize the case where
adjacent classes have the same method, so that we can group the classes
in the dispatch.  That fails for polymorphic methods, which are pretending to
be the same method, but actually do a class test on the receiver internally. In
this case, we should not expand the classes, because it messes up some other
inlining optimizations.

R=vegorov@google.com
BUG=https://github.com/dart-lang/sdk/issues/29892

Review-Url: https://codereview.chromium.org/2956653002 .