VM-codegen: Prefer simpler is-check variant where possible
Before we try anything more complex, check whether there
is only one class that satisfies an 'is' check. If so, we
can just check for that class id and deopt when new classes
are defined.  This change merely reverses the order of
existing code.

This is a reland of https://codereview.chromium.org/2899283002
after an unrelated bug that caused a crash was fixed at
https://github.com/dart-lang/sdk/issues/29774

R=kustermann@google.com, vegorov@google.com
BUG=

Review-Url: https://codereview.chromium.org/2955573002 .