Compute subtyped names.
This is a step to optimizing implemented classes / members feature.

The idea is that we don't need to know elements to make this feature
work. We just need to know whether at least one exists. So, use these
subtyped names to pre-filter files, then read resolved units for special
type of index (to be implemented) with resolved subtyped classes with
overridden class members, then recurse. I think this would be much
faster than to actually resynthesize elements.

R=brianwilkerson@google.com
BUG=

Review-Url: https://codereview.chromium.org/2961773002 .