VM-codegen: Restrict widening class ranges in method dispatch
We widen the range of class-ids that we dispatch to if the
neighbouring classes dispatch to the same methods.  With this
change we don't widen the range to encompass abstract classes
unless we also catch more concrete classes in that way.

This improves the Megamorhphic benchmark suite scores by about
10% on average.

This is split off from https://codereview.chromium.org/2955073005/

R=vegorov@google.com
BUG=

Review-Url: https://codereview.chromium.org/2962333002 .