[profiler, linux/android] Check whether a sample is ARM or Thumb code to decide which FP register to load.
Android and Linux use R11 as the FP register in ARM code and R7 in Thumb code. We've been assuming all code in our process is ARM and reading R11 as FP. This made our stack walks fail if they started in Thumb code.

This issue does not arise on iOS, because its ABI uses R7 as the FP register for both ARM and Thumb code.

R=zra@google.com

Review-Url: https://codereview.chromium.org/2965823002 .