add analysis server --flutter-repo startup flag
This adds a new flag used by `flutter analyze --watch` when analyzing
the flutter repository. When specified, this flag causes analysis server
to enable the public_member_api_docs lint even though it is not enabled
in the analysis_options file. `flutter analyze --watch` needs the results
of this lint to summarize the number of public members missing dartdoc,
but enabling it in the analysis_options causes too much noise in the IDE.

See https://github.com/flutter/flutter/issues/10721 for more background

R=devoncarew@google.com

Review-Url: https://codereview.chromium.org/2976473003 .